# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FabricPro Job Work Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chat/search:
    get:
      operationId: searchChatUsers
      tags: [chat]
      summary: Search users to start a chat with
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Matching users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserPublic"
  /chat/{userId}:
    get:
      operationId: getChatMessages
      tags: [chat]
      summary: Get all messages with a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a message to a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "200":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /chat/{userId}/conversation:
    delete:
      operationId: deleteConversation
      tags: [chat]
      summary: Delete all messages between current user and another user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean

  # --- USERS (for admin) ---
...
components:
  schemas:
...
    Message:
      type: object
      properties:
        id:
          type: integer
        fromUserId:
          type: integer
        toUserId:
          type: integer
        type:
          type: string
          enum: [text, voice, image]
        content:
          type: string
        replyToId:
          type: integer
        replyPreview:
          type: string
        readAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, fromUserId, toUserId, type, content, createdAt]
    SendMessageRequest:
      type: object
      properties:
        type:
          type: string
          enum: [text, voice, image]
        content:
          type: string
        replyToId:
          type: integer
        replyPreview:
          type: string
      required: [type, content]
...
    UserPublic:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        mobile:
          type: string
        avatarUrl:
          type: string
        isOnline:
          type: boolean
        connectionStatus:
          type: string
          enum: [none, pending, accepted]
      required: [id, code]
...